Refactoring code so that the geo-location functionality is seperated from the router
